fixup! refactor integrated code
