+++ sdk/highlight-run/src/index.tsx
@@ -23,10 +23,7 @@ import {
 } from './integrations/mixpanel.js'
 
 import { HIGHLIGHT_URL } from './client/constants/sessions.js'
-import type {
-	ErrorMessageType,
-	Source,
-} from './client/types/shared-types'
+import type { ErrorMessageType, Source } from './client/types/shared-types'
 import {
 	getPreviousSessionData,
 	loadCookieSessionData,
@@ -40,7 +37,7 @@ import { HighlightSegmentMiddleware } from './integrations/segment.js'
 import { initializeFetchListener } from './listeners/fetch'
 import { initializeWebSocketListener } from './listeners/web-socket'
 import { getNoopSpan } from './client/otel/utils.js'
-import { setupLaunchDarklyIntegration } from './client/integrations/launchdarkly.js'
+import { setupLaunchDarklyIntegration } from './integrations/launchdarkly.js'
 import { ViewportResizeListenerArgs } from './client/listeners/viewport-resize-listener'
 
 enum MetricCategory {
